requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.4
Flask==2.0.3
frozenlist==1.3.3
future==0.18.2
gunicorn==20.1.0
idna==3.4
Jinja2==3.0.3
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.2
pandas==1.4.4
pytz==2022.6
requests==2.28.1
tweepy==4.12.1
urllib3==1.26.12
Werkzeug==2.0.3
wincertstore==0.2
yarl==1.8.1